uiApp/Anlzer_Project/requirements.txt
django==1.8.6
requests
django-sass-processor
django-widget-tweaks
django-allauth
djangorestframework
markdown
django-filter
selenium
tweepy
boto3
docker-py
Unidecode

# Scientific Operations
nltk
NumPy==1.11.0
SciPy==0.15.1
scikit-learn